Update dependencies and refactor docker environment variables
Updated the version of 'wbtools' in requirements.txt to keep up with the latest releases. Removed 'TAZENDRA_SSH_USER' and 'TAZENDRA_SSH_PASSWD' environment variables from Dockerfile as they are not needed anymore and moved the environment variable handling to a startup script for a cleaner Dockerfile. This change also makes the codebase more maintainable due to decreased complexity in 'process_papers.py' file and crontab. This is part of an ongoing effort to improve code quality and maintainability.